Switch interpolation to bicubic
