+++ go.mod
@@ -6,10 +6,13 @@ require (
 	dario.cat/mergo v1.0.2
 	github.com/buildkite/terminal-to-html/v3 v3.16.8
 	github.com/clipperhouse/displaywidth v0.11.0
+	github.com/cnkei/gospline v0.0.0-20191204052713-d67fac29a294
 	github.com/fogleman/gg v1.3.0
 	github.com/goastro/twilight v0.0.0-20220106142310-95fee74adefe
+	github.com/gookit/color v1.6.0
 	github.com/hashicorp/golang-lru/v2 v2.0.7
 	github.com/itchyny/gojq v0.12.18
+	github.com/kerrigan29a/drawille-go v0.10.2
 	github.com/kixorz/suncalc v1.0.0
 	github.com/klauspost/lctime v0.1.0
 	github.com/nathan-osman/go-sunrise v1.1.0
@@ -30,7 +33,6 @@ require (
 	github.com/clipperhouse/uax29/v2 v2.7.0 // indirect
 	github.com/davecgh/go-spew v1.1.1 // indirect
 	github.com/golang/freetype v0.0.0-20170609003504-e2365dfdc4a0 // indirect
-	github.com/gookit/color v1.6.0 // indirect
 	github.com/itchyny/timefmt-go v0.1.7 // indirect
 	github.com/mattn/go-runewidth v0.0.19 // indirect
 	github.com/mattn/go-sqlite3 v2.0.3+incompatible // indirect
